+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Features
     * org.springframework.context.annotation.Configuration
     * org.springframework.jmx.export.annotation.ManagedResource
 * Added possibility to specify which annotations mark a class as being _live code_
+* Added possibility to specify a custom XML analyzer treating either an element's text or attribute as a used class
 
 ###[code changes](https://github.com/ImmobilienScout24/deadcode4j/compare/deadcode4j-maven-plugin-1.2.0...master)